requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
requests==2.31.0
PyGithub==2.3.0
google-generativeai==0.5.4
httpx==0.27.0
aiohttp==3.9.5
jira==3.8.0
